Task 14. Balance Factor

+++ 14-binary_tree_balance.c
@@ -12,8 +12,8 @@ int binary_tree_balance(const binary_tree_t *tree)
 
 	if (tree != NULL)
 	{
-		int leftHeight =  binary_tree_height(tree->left);
-		int rightHeight =  binary_tree_height(tree->right);
+		size_t leftHeight =  binary_tree_height(tree->left);
+		size_t rightHeight =  binary_tree_height(tree->right);
 
 		balanceFactor = leftHeight - rightHeight;
 	}
